+++ InvoicesBudgetValidator/Model/Consolidado.cs
@@ -8,9 +8,9 @@ namespace InvoicesBudgetValidator.Model
 {
     class Consolidado
     {
-        public virtual int Id { get; set; }
-        public virtual int Company_Id { get; set; }
-        public virtual int Vendor_Id { get; set; }
+        public virtual long Id { get; set; }
+        public virtual long Company_Id { get; set; }
+        public virtual long Vendor_Id { get; set; }
         public virtual string RFC { get; set; }
         public virtual string Razon_Social { get; set; }
         public virtual decimal Presupuesto { get; set; }
